requirements.txt
paho-mqtt>=1.6.1
sqlalchemy>=2.0.0
google-cloud-bigquery>=3.11.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
prometheus-client>=0.18.0
requests>=2.31.0

